+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `sensor-extraction/ros-to-png`
   - `post-processing/yolo-object-detection`
   - `post-processing/yolop-lane-detection`
+  - `simulations/batch-managed`
+  - `simulations/k8s-managed`
 - made `modules/core/metadata-storage` module generic and bumped CDK version
 - fix code validation script
 - made `ml-training-on-eks` solution generic and organized the solution contents `training-image/mnist`, `ml-training/k8s-managed` into a single dir